+++ cranko-rc-info-v1
[[projects]]
qnames = ["tectonic_xdv", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_errors", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_xetex_format", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_status_base", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_io_base", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_geturl", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_docmodel", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_dep_support", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_cfg_support", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_mac_core", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_bundles", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_bridge_png", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_bridge_icu", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_bridge_graphite2", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_bridge_fontconfig", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_bridge_flate", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_bridge_freetype2", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_bridge_harfbuzz", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_bridge_core", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_xetex_layout", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_pdf_io", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_engine_xetex", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_engine_xdvipdfmx", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_engine_spx2html", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_engine_bibtex", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic", "cargo"]
bump_spec = "minor bump"

+++

+++ crates/engine_bibtex/src/char_info.rs
@@ -94,7 +94,7 @@ const ID_CLASS: [IdClass; 256] = {
     id_class
 };
 
-pub static CHAR_WIDTH: [i32; 256] = {
+pub static CHAR_WIDTH: [i64; 256] = {
     let mut char_width = [0; 256];
 
     // Values taken from the bibtex web2c implementation
